Update error checking in `send_query()`
An exception is raised iff there was an HTTP related error while
sending a command or a response does not look like it is coming from a
Kea Control Agent. All raised exceptions are of type KeaError. Proper
error responses as documented in the API are logged but results in an
empty dictionary being returned (not an exception being thrown).