Add integration tests that check PaloaltoArp running as a job
Simple integration tests are added both for the case where the job's
netbox has a valid key and the ARP set should be updated and for the
case where the job's netbox has only an invalid key and the ARP set
should not changed.

The tests go through the hassle of creating/editing Netboxes and
ManagementProfiles using the SeedDB endpoints, and will be able to
pick up bugs caused by changes in the naming scheme of the values and
keys in a HTTP_REST_API ManagementProfile instance's configuration
json (e.g. the event that the HTTPRestForm Django form class is
changed so that user-supplied values suddenly are written to different
keys in the resulting ManagementProfile instance's configuration json
than the PaloaltoArp plugin expects).