Fix paloaltoarp's usage of Netbox to work with shadowed netboxes
Shadowed netboxes (ipdevpoll.shadows.Netbox instances as  opposed to
models.manage.Netbox instances) lacks some functionality that the
paloaltoplugin relied on, namely the ability to look at foreign
relationship sets, and the usage of some methods that isn't available
in the shadowed counterpart. Since the netbox instance supplied to the
plugin is the shadowed version, we must hence find alternative ways to
realise this functionality; this is basically done by using the
shadowed netbox's id to look up the actual netbox in the
models.manage.Netbox query_set when needed.